chore: Move the set of current page after the success of load more
We are making this change to address a possible bug where we where increasing the page number,\
in the beginning of the function. This could cause issues where the server returned an error or the request failed for any reason.
Now we are increasing the page after the success of the request.

We have also added a comment regarding the setHasMore logic to provide more clarity around it.