chore: Navigate to cats-gallery instead of root
We are making a change where in some components we where navigating to cats-gallery instead  the root of the application "/".
This is not breaking change, as the "/" is equal to "cats-gallery" as we have modified nextjs to treat as root the "cats-gallery".
The reason of this change is to make it more obvious to the reviewer that the navigation leads to cats-gallery.